requirements.txt
dnspython==2.6.1
Faker==28.4.1
psycopg2-binary==2.9.9
pymongo==4.8.0
python-dateutil==2.9.0.post0
six==1.16.0
